fix(suse): Avoid refresh attempts for Zypper queries
On invocation, these query commands will attempt to refresh repositories.
Since they lack root privileges, they will not be able to save new
package information on the device, but can take lots of time if many
repositories are enabled. For a series of consecutive query commands,
a user has to wait for the pointless 'repository refresh' every time.

This patch asks `zypper` to not refresh repositories. If a user intends
to do that, they would need to use `sudo` anyway, which the `zref` alias
already does.