feat(zbell): option not to use notify-send
When using a workstation remotely with something like SSH, we might not
want to send popup's.

In my case, when I open the screen, I can see heaps of missed
notifications. All I want is the 'bell' sent in the terminal.

Now we have a simple option not to use 'notify-send' even if it is
available.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>