check "diff --color" shouldn't hang the script
The current implementation to check if diff should alias to "diff --color" is
to try if "diff --color . ." works. However when "." cannot diff fast enough,
say if it's a huge, deep directory or if it's a remotely mounted directory, it 
could hang and wait until `diff` completely scans the current directory.

Since all we need is to see if `diff --color` works this commit simply replace
"." with "$0", which is the path to this script and therefore the existence is
guaranteed and can be processed rather quickly.