fix(plugins/jump)! jump to the link, not absolute, path
BREAKING CHANGE

jumping previously took you to the absolute path of the marked
directory. If you create a mark to a symlinked path, then when you jump
you'll end up at the absolute version of that path:

$ mkdir /real-path
$ ln -s /real-path /link-path
$ cd /link-path
$ mark
Mark /link-path as link-path? (y/n) y
$ jump link-path
$ pwd
/real-path

this seems unintuitive to me, so this commit changes the jump function
to take you to the actual path of the mark (`/link-path` in the example
above) rather than the absolute path.