fix(plugin/jump): fix text alignment when printing marks
the `marks` function right-justifies mark names, to make the output
easier to read. However, because of the escape sequences for colors in
the output, the printf command doesn't calculate the padding correctly.
This results in the mark names not actually being right-justified
correctly, with short names not being padded enough. E.g.:

 code -> /home/ditzy/code
 misc -> /home/ditzy/misc
sys-conf -> /home/ditzy/.sys-conf

this commit fixes this problem by only applying padding to the actual
mark name itself, and not the name+colors. E.g.:

    code -> /home/ditzy/code
    misc -> /home/ditzy/misc
sys-conf -> /home/ditzy/.sys-conf