fix(omz): make omz honour symlinks when changing .zshrc (#11082)
For people that version control their dotfiles, ~/.zshrc may be a
symlink. The omz tool was breaking symlinks when updating ~/.zshrc for
theme and plugin operations, thus not making the changes in a way that
are easily preserved and managed.

To fix this, rely on a few globbing tricks when finding the file to
resolve the target for ~/.zshrc when backing it up and when
overwriting it.

This will not work on systems lacking a realpath(3) call, but should
be no worse the current behaviour.