fix(git-auto-fetch): Prevent User/Password Prompts
If the remote is using HTTP authentication (and does not use a
credential helper) it triggers writes to the terminal that cannot be
responded to. This happens every time the repository is fetched by
git-auto-fetch.

`GIT_TERMINAL_PROMPT=0` prevents prompts to the terminal. The man page
explicitly names HTTP authentication as an example.