fix(gh): only create completion cache when needed
Completions were generated and cached on every shell opened.
This fix caches the version the cache was last generated for and
only runs the completion generator in case there is a mismatch or
no completion cache file available.

Also fixed the readme, to state correct path for completions cache and
version cached file.