refactor(archlinux): refactor `pacdisowned` in plugin `archlinux`
In the Archlinux environment, `mktemp` is part of the base installation, so there is no need to worry about its existence.
Remove search path `/bin`, `/sbin` and `/lib`
In the Archlinux environment, `/bin`, `/sbin`, `/lib` and `/lib64` (which even not mentioned in the current version) are just soft links to the folders under /usr.
The command string passed to the `trap` command should use double quotes to ensure that parameter expansion inside is done immediately (tmp_dir is declared as local).
Temporary files should be cleaned up as soon as possible after use, instead of waiting for the `trap` to take effect.
> $ ls -ld /bin /sbin /lib /lib64
> lrwxrwxrwx 1 root root 7 Dec  7 10:41 /bin -> usr/bin
> lrwxrwxrwx 1 root root 7 Dec  7 10:41 /lib -> usr/lib
> lrwxrwxrwx 1 root root 7 Dec  7 10:41 /lib64 -> usr/lib
> lrwxrwxrwx 1 root root 7 Dec  7 10:41 /sbin -> usr/bin