Sort exps by cheatsheet group key to fix grouping
The itertools.groupby function operates by grouping together
*contiguous* items that have the same key. Since the `exps` list is
unordered with respect to the 3rd element (i.e., they key for the
group), this results in chunking together of items with the same key
only if it is fortunate enough to be immediately before or immediately
after another expression with the same key. Without sorting the list of
`exps`, numerous expressions get put into a bin of one when they should
be included in a larger group, which results in erroneously putting
these expressions into the "_default" group.