fix(vi-mode): fix check for prompt redisplay on mode check
The previous logic checked $VI_MODE_RESET_PROMPT_ON_MODE_CHANGE,
which was changed on the `vi_mode_prompt_info` function to make
sure that we redisplay the prompt on vi mode change only if the
vi-mode indicator is shown on the prompt.

This failed because the variable context inside a prompt function
did not propagate to the regular variable environment, so the
`VI_MODE_RESET_PROMPT_ON_MODE_CHANGE=true` code did not actually
change anything and the prompt was never forcibly redisplayed.

This change dynamically checks whether the vi_mode_prompt_info
function is added to the prompt segments to figure out whether
we must redisplay the prompt.