Fixing slowness in large GIT repo (#12088)
Every time you press Enter, we actually execute a git status command
to check if the current directory is dirty. In large Git repositories,
this process can be particularly slow. Therefore, before executing this
command, we check if the repository is larger than 100MB and skip the
command if it is.