feat(aws): Add aws sso logout possibility for aws plugin
The aws plugin gives the user the possibility to login into the sso
session when using the AWS profile selection. Unfortunately, it is not
possible to also logout this way. Naturally, one would think that when
one is typing "asp <aws_profile> login" it is also possible to type
"asp <aws_profile> logout" but this is currently not the case. This
commit adds this possibility and therefore enables the user to login and
logout with the plugin without having to use awscli directly.