fix(bun): bun completions should explicitly set SHELL
`bun completions` results in a file according to the current $SHELL, but if you don´t have ZSH as your $SHELL, the plugin will create a bash script as the completions script. This commit fixes that case.

Users with broken completions probably need to remove their `$ZSH_CACHE_DIR/completions/_bun` so it is created again.