Remove last zsh call in install.sh
The call restarts zsh with the `-l` flag. In my case, I'm writing a script to bootstrap a new machine and this forces me to either put the ohmyzsh installation at the end, or to find some other way.

I would like to either choose by a flag to run that, for now that's just a convenience feature.