Remove echo from gradle plugin
When using this plugin and running e.g. `gradle --quiet`, the echo statement would print to stdout. I figured that I could make it respect the `--quiet` flag (and possibly various other situations in which echoing isn't expected to print to stdout), but as a much simpler approach it can also be removed altogether unless there is a reason to keep it.