Fixed poetry-env plugin for Poetry >=2.0.0
Since version 2.0.0 poetry stopped using '[project.tool]' section in `pyproject.toml`. So the plugin didn't work for the newer poetry version.

With this change, the plugin will check for `pyproject.toml` and `poetry.lock` files to reliably check if it's a poetry project. In my testing, it worked with both poetry versions 1.8.5 and 2.0.1.