fix(bgnotify): detect if sway is running and not just installed
Before it also tried to call swaymsg if running on another wayland
compositor, but sway was also installed. The fix is to also check for
the $SWAYSOCK variable, which swaymsg is using to check where to talk to
sway. So it only tries to run swaymsg if sway is actually running.