fix(poetry-env): fix env detection for pyproject without [tool.poetry]
Since version 2.0.0, poetry stopped requiring '[tool.poetry]' section in pyproject.toml and new environments are created without it, but the zsh plugin still required it.

With this change, the plugin will check for `pyproject.toml` and `poetry.lock` files to reliably check if it's a poetry project. Tested with both poetry versions 1.8.5, 2.0.1 and 2.1.2.