fix(plugins/git): change alias `gm` to `gme` (conflict with GraphicsMagick)
`gm` is the name of the GraphicsMagick (http://www.graphicsmagick.org/index.html) command line tool, which conflicts with the `gm` alias in the git plugin. This change renames the alias to `gme` to avoid the conflict.

style(plugins/git): realign column in Readme

fix(plugins/git): realign column in Readme