feat(systemd): preserve editor env for sudo commands
When running `systemctl` commands that require sudo, the users'
configured editor environment variables were not preserved. This
resulted in systemctl opening a default editor.

By instructing sudo to preserve the common environment variables for
configuring an editor, systemctl can see these variables and open the
users' preferred editor.

This is mostly relevant for the `sc-edit` alias.