fix(git): git_status_prompt must respects spaces to properly match prefixes
`git_status_prompt` was always empty as whitespace around the keys for `$prefix_constant_map` were trimmed. The in list of the for loop must be enclosed in `"` to avoid trimming, which in return requires the iterator specifier to contain an `@` to avoid concatenation of all keys. This is important for the regex to work properly.